+++ lib/src/dataservices/interfaces.ts
@@ -3,7 +3,7 @@ export type HttpMethods = 'DELETE' | 'GET' | 'POST' | 'PUT';
 export interface RequestData {
   method: HttpMethods;
   url: string;
-  options: any;
+  options?: any;
 }
 
 /**
@@ -11,4 +11,6 @@ export interface RequestData {
  * Same as HttpClient's HttpParamsOptions which is NOT exported at package level
  * https://github.com/angular/angular/issues/22013
  */
-export interface QueryParams { [name: string]: string | string[]; }
+export interface QueryParams {
+  [name: string]: string | string[];
+}